pyproject.toml
[project]
name = "ad_rag_app"
version = "0.1.0"
description = "Alzheimer’s Disease RAG Microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.84",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
filterwarnings = [
  "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"